Release v0.8.2: Enhanced adaptive learning performance optimization system
- Add comprehensive adaptive_learning_analyzer tool for performance data analysis
- Enhance parallel execution with improved algorithmic threshold detection
- Update version numbers in CMakeLists.txt and libstats.h to 0.8.2
- Improve documentation and integration across all performance components
- Add robust error handling and validation throughout the system
- Optimize memory management and resource utilization
- Enhance test coverage and validation frameworks

Key features in v0.8.2:
* Advanced adaptive learning analyzer tool with detailed performance metrics
* Improved parallel execution policies and threshold optimization
* Enhanced system capability detection and benchmarking
* Better integration between performance components
* Comprehensive performance data collection and analysis
* Robust error handling and resource management

+++ examples/performance_learning_demo.cpp
@@ -219,13 +219,13 @@ void demonstrate_adaptive_learning() {
             // Test multiple strategies for comparison
             for (auto strategy : test_strategies) {
                 // Add some realistic variance based on strategy (simulated)
-                uint64_t adjusted_duration = duration;
+                uint64_t adjusted_duration = static_cast<uint64_t>(duration);
                 switch (strategy) {
                     case libstats::performance::Strategy::SCALAR:
                         adjusted_duration = static_cast<uint64_t>(static_cast<double>(duration) * 1.5); // Slower
                         break;
                     case libstats::performance::Strategy::SIMD_BATCH:
-                        adjusted_duration = duration; // Baseline
+                        adjusted_duration = static_cast<uint64_t>(duration); // Baseline
                         break;
                     case libstats::performance::Strategy::PARALLEL_SIMD:
                         if (size > 5000) {
